+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 6.0.0
 - *Breaking change* Angular is upgraded to version 8.
 - *Breaking change* `title-value` is now declarative using: `title-value-heading` and `title-value-container` as children.
+- *Deprecated* `list` and `list-item`.
 - Fixed `title-value-heading` and `title-value-container` now has a new property for defining it's width.
-- Added width-handler for `vgr-datepicker` & `vgr-input`
-- Deprecated `list` and `list-item`
+- Fixed `vgr-dropdown-select` uses `vgr-filter-textbox` for filtering.
+- Added width-handler for `vgr-datepicker`, `vgr-input` & `vgr-dropdown-select`
 
 ## 5.1.3 (Stable, 2019-10-21)
 - Fixed another bug in "grid"-component regarding toggle animation.